Warning fixes

+++ programs/Xserver/hw/xfree86/drivers/i740/i740.h
@@ -25,7 +25,7 @@ TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 **************************************************************************/
-/* $XFree86: xc/programs/Xserver/hw/xfree86/drivers/i740/i740.h,v 1.5 2002/01/25 21:56:02 tsi Exp $ */
+/* $XFree86: xc/programs/Xserver/hw/xfree86/drivers/i740/i740.h,v 1.6tsi Exp $ */
 
 /*
  * Authors:
@@ -95,8 +95,8 @@ typedef struct _I740Rec {
   int MaxClock;
   int CursorStart;
   int Chipset;
-  int LinearAddr;
-  int MMIOAddr;
+  unsigned long LinearAddr;
+  unsigned long MMIOAddr;
   EntityInfoPtr pEnt;
   pciVideoPtr PciInfo;
   PCITAG PciTag;
